# feat(ige): reduce allocations for DecryptBlocks/EncryptBlocks