--- a/encrypt_test.go
+++ b/encrypt_test.go
@@ -3,6 +3,7 @@ package ige
 import (
 	"bytes"
 	"crypto/aes"
+	"fmt"
 	"testing"
 )
 
@@ -82,4 +83,39 @@ func TestEncryptCryptBlocksPanicDst(t *testing.T) {
 
 	i := NewIGEEncrypter(c, make([]byte, 32))
 	i.CryptBlocks(make([]byte, 1), make([]byte, 16))
+}
+
+func BenchmarkEncryptBlocks(b *testing.B) {
+	for _, payload := range []int{
+		16,
+		128,
+		1024,
+		8192,
+		64 * 1024,
+		512 * 1024,
+	} {
+		b.Run(fmt.Sprintf("%d", payload), benchEncrypt(payload))
+	}
+}
+
+func benchEncrypt(n int) func(b *testing.B) {
+	return func(b *testing.B) {
+		b.Helper()
+
+		src := make([]byte, n)
+		dst := make([]byte, n)
+
+		b.ReportAllocs()
+		b.SetBytes(int64(n))
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			c, err := aes.NewCipher(TestVectors[0].Key)
+			if err != nil {
+				b.Fatal(err)
+			}
+
+			EncryptBlocks(c, TestVectors[0].IV, dst, src)
+		}
+	}
 }